requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
python-dotenv
pdfplumber
PyPDF2
tabula-py
pandas
google-generativeai
anthropic
openai
chromadb
sentence-transformers
langchain
langchain-community
requests
httpx
aiohttp
numpy
beautifulsoup4
pydantic
python-dateutil
feedparser
